refactor: 재료충전 분리
