+++ src/components/CoffeeMachineComponent.js
@@ -8,13 +8,14 @@ import {
   validateCafeLatteMaterialQuantity,
   validateChargeCoinInput,
   validateCoffeeMaterialQuantity,
-  validateMaterialInput,
   validateMilkMaterialQuantity,
 } from '../utils/validations';
+import RechargeMaterialComponent from './RechargeMaterialComponent';
 
 class CoffeeMachineComponent {
   constructor() {
     this.initDOM();
+    this.rechargeComponent = new RechargeMaterialComponent();
     this.showPurchaseCoffeeComponent();
     this.bindEventListener();
     this.drink = new Drink();
@@ -31,11 +32,6 @@ class CoffeeMachineComponent {
     this.$purchasableDrinkQuantity = document.querySelectorAll('.drink-quantity');
     this.$purchaseButtons = document.querySelectorAll('.purchase-button');
 
-    this.$coffeeBeanQuantityElement = document.querySelector('#coffee-beans-quantity');
-    this.$cupQuantityElement = document.querySelector('#cups-quantity');
-    this.$milkQuantityElement = document.querySelector('#milk-quantity');
-
-    this.$rechargeDrinkButtonContainer = document.querySelector('.recharge-drink-container');
     this.$totalChargeCoinElement = document.querySelector('#total-charge-coin');
     this.$chargeCoinButton = document.querySelector('#charge-coin-submit');
   }
@@ -48,18 +44,6 @@ class CoffeeMachineComponent {
       this.$purchasableDrinkQuantity.forEach((item, index) => {
         item.textContent = purchaseDrinkQuantity[index];
       });
-    }
-  }
-
-  showNowMaterialQuantity() {
-    const materials = materialStore.getMaterialStore();
-
-    if (materials !== 0) {
-      const { coffeeBean, cup, milk } = materials;
-
-      this.$coffeeBeanQuantityElement.textContent = coffeeBean;
-      this.$cupQuantityElement.textContent = cup;
-      this.$milkQuantityElement.textContent = milk;
     }
   }
 
@@ -81,13 +65,12 @@ class CoffeeMachineComponent {
     this.$rechargeTab.classList.add('is-active');
     this.$purchaseTab.classList.remove('is-active');
 
-    this.showNowMaterialQuantity();
+    this.rechargeComponent.showNowMaterialQuantity();
   }
 
   bindEventListener() {
     this.$nav.addEventListener('click', this.onNavButtonClick);
     this.$purchaseDrinkButtonContainer.addEventListener('click', this.onPurchaseDrinkButtonClick);
-    this.$rechargeDrinkButtonContainer.addEventListener('click', this.onRechargeButtonClick);
     this.$chargeCoinButton.addEventListener('click', this.onChargeCoinButtonClick);
   }
 
@@ -153,58 +136,6 @@ class CoffeeMachineComponent {
     this.showPurchasableDrinkQuantity();
     this.activePurchaseMenuButton();
     this.showTotalChargeCoin();
-  };
-
-  onRechargeButtonClick = e => {
-    e.preventDefault();
-    if (e.target.id === 'recharge-coffee-beans-button') {
-      this.onRechargeCoffeeBeanButtonClick();
-    }
-    if (e.target.id === 'recharge-milk-button') {
-      this.onRechargeMilkButtonClick();
-    }
-    if (e.target.id === 'recharge-cups-button') {
-      this.onRechargeCupButtonClick();
-    }
-    this.showNowMaterialQuantity();
-  };
-
-  onRechargeCoffeeBeanButtonClick = () => {
-    const $rechargeCoffeeBeanInput = document.querySelector('#recharge-coffee-beans-input');
-    const { valueAsNumber: coffeeBeanInputValue } = $rechargeCoffeeBeanInput;
-    $rechargeCoffeeBeanInput.value = '';
-
-    if (!validateMaterialInput(coffeeBeanInputValue)) {
-      showSnackBar(ERROR_MSG.INVALID_QUANTITY_INPUT);
-      return;
-    }
-
-    materialStore.rechargeCoffeeBean(coffeeBeanInputValue);
-  };
-
-  onRechargeMilkButtonClick = () => {
-    const $rechargeMilkInput = document.querySelector('#recharge-milk-input');
-    const { valueAsNumber: milkInputValue } = $rechargeMilkInput;
-    $rechargeMilkInput.value = '';
-
-    if (!validateMaterialInput(milkInputValue)) {
-      showSnackBar(ERROR_MSG.INVALID_QUANTITY_INPUT);
-      return;
-    }
-    materialStore.rechargeMilk(milkInputValue);
-  };
-
-  onRechargeCupButtonClick = () => {
-    const $rechargeCupInput = document.querySelector('#recharge-cups-input');
-    const { valueAsNumber: cupInputValue } = $rechargeCupInput;
-    $rechargeCupInput.value = '';
-
-    if (!validateMaterialInput(cupInputValue)) {
-      showSnackBar(ERROR_MSG.INVALID_QUANTITY_INPUT);
-      return;
-    }
-
-    materialStore.rechargeCup(cupInputValue);
   };
 
   onChargeCoinButtonClick = e => {
